dealing with test weirdness

+++ crates/nu_plugin_polars/src/dataframe/command/data/struct_json_encode.rs
@@ -31,7 +31,7 @@ impl PluginCommand for StructJsonEncode {
         vec![Example {
             description: "Encode a struct as JSON",
             example: r#"[[id person]; [1 {name: "Bob", age: 36}] [2 {name: "Betty", age: 63}]] 
-                    | polars into-df -s {id: i64, person: {name: str, age: u8}} 
+                    | polars into-df -s {id: i32, person: {name: str, age: u8}} 
                     | polars select id (polars col person | polars struct-json-encode | polars as encoded)
                     | polars sort-by id
                     | polars collect"#,
